refactor(daemon): organize dispatch and spawner tests
